+++ src/modules/burpread_mod.f90
@@ -1616,10 +1616,12 @@ CONTAINS
     real   , ALLOCATABLE   :: OBSERV  (:,:),    OBSERV_SFC(:,:)
     real   , ALLOCATABLE   :: BiasCorrection_sfc(:,:,:)
     real   , ALLOCATABLE   :: BCOR_SFC(:,:)
+    INTEGER, PARAMETER     :: MAXRONVAL=500
+    real                   :: ROLAT(MAXRONVAL), ROLON(MAXRONVAL)
 
     INTEGER, ALLOCATABLE   :: MTVAL(:)
     INTEGER, ALLOCATABLE   :: HAVAL(:), GAVAL(:), QI1VAL(:) ,QI2VAL(:), LSVAL(:)
-    REAL(pre_obsReal) , ALLOCATABLE  :: azimuth(:)
+    real(pre_obsReal) , ALLOCATABLE  :: azimuth(:)
     INTEGER, ALLOCATABLE   :: QCFLAG  (:,:,:),  QCFLAG_SFC(:,:,:)
     INTEGER, ALLOCATABLE   :: QCFLAGS (:,:),   QCFLAGS_SFC(:,:)
     integer, allocatable   :: hiresTimeFlag(:,:), hiresLatFlag(:,:)
@@ -1651,22 +1653,22 @@ CONTAINS
     integer                :: NBELE,NVALE,NTE
     integer                :: J,JJ,K,KK,KL,IL,ERROR,OBSN
     integer                :: info_elepos,IND_ELE,IND_VCOORD,IND_QCFLAG,IND_SW
-    integer                :: IND055200,IND4208,ind4197,IND5002,IND6002,ind_al
+    integer                :: IND055200,IND4208,ind4197,IND5002,IND6002,ind_al,IND5001,IND6001
     integer                :: IND_LAT,IND_LON,IND_TIME,IND_EMIS,IND_BCOR,IND_PHASE,IND_BCOR_TT,IND_BCOR_HU
     integer                :: FLAG_PASSAGE1,FLAG_PASSAGE2,FLAG_PASSAGE3,FLAG_PASSAGE4
     integer                :: IND_dataQcFlag0, IND_dataQcFlag1, IND_dataQcFlag2 
 
     integer                :: vcord_type(10),SUM,vcoord_type
     REAL(pre_obsReal)      :: RELEV,XLAT,XLON,RELEV2
-    real                   :: XTIME
-    integer                :: status ,idtyp,lati,long,dx,dy,elev, &
-                               drnd,date_h,hhmm_h,oars,runn,YMD_DATE,HM,kstamp,kstamp2,HM_SFC,YMD_DATE_SFC
+    real                   :: XTIME,ROLAT0,ROLON0,ROLAT1,ROLON1
+    integer                :: status ,idtyp,lati,long,dx,dy,elev
+    integer                :: drnd,date_h,hhmm_h,oars,runn,YMD_DATE,HM,kstamp,kstamp2,HM_SFC,YMD_DATE_SFC
 
     integer                :: iele,NELE,NELE_SFC,NVAL,NT,NELE_INFO,LN
     integer                :: bit_alt,btyp_offset,btyp_offset_uni
     character(len = 5)     :: BURP_TYP
     CHARACTER(LEN=9)       :: STNID,STN_RESUME
-    LOGICAL                :: HIRES,HIRES_SFC,HIPCS,phasePresent
+    LOGICAL                :: HIRES,HIRES_SFC,HIPCS,phasePresent,LOK,LROK
     integer                :: NDATA,NDATA_SF
     integer                :: IER,date2,time2,time_sonde,NEWDATE
     real                   :: RAD_MOY,RAD_STD
@@ -2020,6 +2022,8 @@ CONTAINS
         HIPCS=.FALSE.
         HIRES_SFC=.FALSE.
         phasePresent = .false.
+        LROK = .false.
+
         BLOCKS1: do
 
           ref_blk = BURP_Find_Block(Rpt_in, &
@@ -2037,6 +2041,36 @@ CONTAINS
                       & BTYP   = btyp, &
                       & BKSTP  = BKSTP, &
                       & IOSTAT = error)
+
+          ! Read slant latlon if type is RO
+          if (trim(familytype) == 'RO' .and. bfam == 0 .and. LROK == .FALSE.) then
+            ROLAT0 = 0.01*lati- 90.
+            ROLON0 = 0.01*long
+            if (ROLON0 > 180.) ROLON0 = ROLON0-360.
+            ROLAT(:) = ROLAT0
+            ROLON(:) = ROLON0
+            IND5001 = BURP_Find_Element(Block_in, ELEMENT = 5001, IOSTAT = error)
+            IND6001 = BURP_Find_Element(Block_in, ELEMENT = 6001, IOSTAT = error)
+            if (IND5001 > 0 .and. IND6001 > 0) then
+              do j = 1, nvale
+                ROLAT1 = BURP_Get_Rval(Block_in, &
+                                       NELE_IND = IND5001, &
+                                       NVAL_IND = j, &
+                                       NT_IND = 1, IOSTAT = error)
+                ROLON1 = BURP_Get_Rval(Block_in, &
+                                       NELE_IND = IND6001, &
+                                       NVAL_IND = j, &
+                                       NT_IND = 1, IOSTAT = error)
+                lok = ( -90.1 < ROLAT1 .and. ROLAT1 <  90.1) .and. &
+                      (-180.1 < ROLON1 .and. ROLON1 < 360.1)
+                if (lok .and. j<=MAXRONVAL) then
+                  ROLAT(j) = ROLAT1
+                  ROLON(j) = ROLON1
+                  LROK = .TRUE.
+                end if
+              end do
+            end if
+          end if
 
           ! observation block (btyp = 0100 100011X XXXX)
           if(trim(familytype) == 'AL')then
@@ -2651,7 +2685,7 @@ CONTAINS
 
               NDATA_SF= WRITE_BODY(obsdat,UNI_FAMILYTYPE,RELEV,vcoord_sfc,vcoord_type, &
                                    OBSERV_SFC,qcflags_sfc,NELE_SFC,1,LISTE_ELE_SFC, &
-                                   dataQcFlagLEV_sfc, BiasCorrection_opt=BCOR_SFC)              
+                                   dataQcFlagLEV_sfc,ROLAT,ROLON,BiasCorrection_opt=BCOR_SFC)
 
               IF ( NDATA_SF > 0) THEN
                 call WRITE_HEADER(obsdat,STNID,XLAT,XLON,YMD_DATE_SFC,HM_SFC,idtyp,STATUS,RELEV,FILENUMB)
@@ -2706,14 +2740,19 @@ CONTAINS
 
                 VCORD(1)=VCOORD(jj,k)
                 if (allocated(BCOR)) then
-                  NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV,qcflags,NELE,1,LISTE_ELE, &
-                                    dataQcFlagLEV, SURF_EMIS_opt = SURF_EMIS, BiasCorrection_opt = BiasCorrection)
+                  NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type, &
+                                    OBSERV,qcflags,NELE,1,LISTE_ELE,dataQcFlagLEV,ROLAT,ROLON, &
+                                    SURF_EMIS_opt=SURF_EMIS, &
+                                    BiasCorrection_opt=BiasCorrection)
                 elseif (allocated(BCOR2)) then
-                  NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV,qcflags,NELE,1,LISTE_ELE, &
-                                    dataQcFlagLEV, SURF_EMIS_opt = SURF_EMIS, BiasCorrection_opt = BiasCorrection2)
+                  NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type, &
+                                    OBSERV,qcflags,NELE,1,LISTE_ELE,dataQcFlagLEV,ROLAT,ROLON, &
+                                    SURF_EMIS_opt=SURF_EMIS, &
+                                    BiasCorrection_opt=BiasCorrection2)
                 else
-                  NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV,qcflags,NELE,1,LISTE_ELE, &
-                                    dataQcFlagLEV, SURF_EMIS_opt = SURF_EMIS)
+                  NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type, &
+                                    OBSERV,qcflags,NELE,1,LISTE_ELE,dataQcFlagLEV,ROLAT,ROLON, &
+                                    SURF_EMIS_opt=SURF_EMIS)
                 end if
 
                 IF (NDATA > 0) THEN
@@ -2788,8 +2827,10 @@ CONTAINS
               ! dataQcFlagLev does not exist for surface data
               dataQcFlagLev_sfc(:) = MPC_missingValue_INT
 
-              NDATA_SF= WRITE_BODY(obsdat,UNI_FAMILYTYPE,RELEV,vcoord_sfc,vcoord_type,OBSERV_sfc,qcflags_sfc, &
-                                   NELE_SFC,1,LISTE_ELE_SFC, dataQcFlagLEV_sfc, BiasCorrection_opt = BCOR_SFC)
+              NDATA_SF= WRITE_BODY(obsdat,UNI_FAMILYTYPE,RELEV,vcoord_sfc,vcoord_type, &
+                                   OBSERV_sfc,qcflags_sfc,NELE_SFC,1,LISTE_ELE_SFC, &
+                                   dataQcFlagLEV_sfc,ROLAT,ROLON,BiasCorrection_opt=BCOR_SFC)
+
               IF ( NDATA_SF > 0) THEN
                 call WRITE_HEADER(obsdat,STNID,XLAT,XLON,YMD_DATE,HM,idtyp,STATUS,RELEV,FILENUMB)
                 OBSN=obs_numHeader(obsdat) 
@@ -2824,14 +2865,17 @@ CONTAINS
 
               !CASES DEPENDING ON WETHER ON NOT WE HAVE MW DATA
               if (allocated(BCOR)) then
-                NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV,qcflags,NELE,NVAL,LISTE_ELE, &
-                                  dataQcFlagLEV, SURF_EMIS_opt = SURF_EMIS, BiasCorrection_opt = BiasCorrection)
+                NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV, &
+                                  qcflags,NELE,NVAL,LISTE_ELE,dataQcFlagLEV,ROLAT,ROLON, &
+                                  SURF_EMIS_opt=SURF_EMIS,BiasCorrection_opt=BiasCorrection)
               elseif (allocated(BCOR2)) then
-                NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV,qcflags,NELE,NVAL,LISTE_ELE, &
-                                  dataQcFlagLEV, SURF_EMIS_opt = SURF_EMIS, BiasCorrection_opt = BiasCorrection2)
+                NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV, &
+                                  qcflags,NELE,NVAL,LISTE_ELE,dataQcFlagLEV,ROLAT,ROLON, &
+                                  SURF_EMIS_opt=SURF_EMIS,BiasCorrection_opt=BiasCorrection2)
               else
-                NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV,qcflags,NELE,NVAL,LISTE_ELE, &
-                                  dataQcFlagLEV, SURF_EMIS_opt = SURF_EMIS)
+                NDATA= WRITE_BODY(obsdat,familytype,RELEV,VCORD,vcoord_type,OBSERV, &
+                                  qcflags,NELE,NVAL,LISTE_ELE,dataQcFlagLEV,ROLAT,ROLON, &
+                                  SURF_EMIS_opt=SURF_EMIS)
               end if
               
               IF (NDATA > 0) THEN
@@ -3035,32 +3079,30 @@ CONTAINS
 
 
   FUNCTION WRITE_BODY(obsdat,FAMTYP, ELEV,VERTCOORD,VCOORD_TYPE, &
-                      obsvalue,qcflag,NELE,NVAL,LISTE_ELE,dataQcFlagLEV, SURF_EMIS_opt, &
-                      BiasCorrection_opt)
+                      obsvalue,qcflag,NELE,NVAL,LISTE_ELE,dataQcFlagLEV,ROLAT,ROLON, &
+                      SURF_EMIS_opt,BiasCorrection_opt)
 
     implicit none
 
+    ! Arguments:
     type (struct_obs), intent(inout) :: obsdat
-    integer ::  WRITE_BODY,VCOORD_TYPE
-    real   , allocatable          ::  OBSVALUE(:,:)
-    real   , allocatable,optional ::  SURF_EMIS_opt(:)
+    INTEGER ::  WRITE_BODY,VCOORD_TYPE
+    REAL   , allocatable          ::  OBSVALUE(:,:)
     INTEGER, allocatable          ::  QCFLAG(:,:)
-    real   , allocatable          ::  VERTCOORD(:)
-    real   , allocatable,optional ::  BiasCorrection_opt(:,:)
+    REAL   , allocatable          ::  VERTCOORD(:)
+    REAL                          ::  ROLAT(:), ROLON(:)
+    REAL   , allocatable,optional ::  SURF_EMIS_opt(:)
+    REAL   , allocatable,optional ::  BiasCorrection_opt(:,:)
     integer, intent(in)           ::  dataQcFlagLEV(:)
-    CHARACTER*2 ::   FAMTYP
-    real        ::   ELEVFACT,VCOORD
-    integer     ::   NELE,NVAL
-    integer     ::   LISTE_ELE(:)
-    integer     ::   NOBS,VARNO,IL,J,COUNT,NLV
-    integer     ::   IFLAG,BITSflagoff,BITSflagon,IFLAG2
-    REAL(pre_obsReal) :: MISG,OBSV,ELEV,ELEV_R,REMIS,emmissivite,BCOR
-    integer     ::   VCO,NONELEV
-    real        ::   ZEMFACT
-    LOGICAL     ::   L_EMISS
-    LOGICAL     ::   L_BCOR
-    LOGICAL     ::   L_dataQcFlag2
 
+    ! Locals:
+    CHARACTER(len=2)  :: FAMTYP
+    REAL              :: ELEVFACT,VCOORD,ZEMFACT
+    INTEGER           :: NELE,NVAL,VCO,NONELEV
+    integer           :: LISTE_ELE(:),ID_OBS,NOBS,VARNO,IL,J,COUNT,NLV
+    INTEGER           :: IFLAG,IFLAG2,BITSflagoff,BITSflagon
+    REAL(pre_obsReal) :: MISG,OBSV,ELEV,ELEV_R,REMIS,emmissivite,BCOR,rolat1,rolon1
+    LOGICAL           :: L_EMISS,L_BCOR,L_dataQcFlag2
     
     L_EMISS = present( SURF_EMIS_opt )
     L_BCOR  = present( BiasCorrection_opt )
@@ -3186,7 +3228,12 @@ CONTAINS
           ELEV_R=VCOORD + ELEV*ELEVFACT
           call obs_bodySet_r(obsdat,OBS_PPP,count, ELEV_R)
           call obs_bodySet_i(obsdat,OBS_FLG,count,IFLAG)
-
+          if ( FAMTYP == 'RO' ) then
+            rolat1 = rolat(j)*MPC_RADIANS_PER_DEGREE_R8
+            rolon1 = rolon(j)*MPC_RADIANS_PER_DEGREE_R8
+            call obs_bodySet_r(obsdat,OBS_ROLA,count,rolat1)
+            call obs_bodySet_r(obsdat,OBS_ROLO,count,rolon1)
+          end if
           if ( L_BCOR .and. obs_columnActive_RB(obsdat,OBS_BCOR) ) then
             call obs_bodySet_r(obsdat,OBS_BCOR,count,BCOR)
           end if