+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+ * Added ability to use slant path for GPS-RO observations (#380 and !374)
  * Added quality control for AMSUA and ATMS to `obsSelection` (#335 and !320)
  * Added interpolation from a `Y` grid for the background fields (#385 and !367)
  * When `midas.splitobs` is splitting an SQLite observation file into